requirements.txt
asgiref==3.8.1
Django==5.0.4
django-ckeditor-5==0.2.12
django-multiselectfield==0.1.12
djangorestframework==3.15.1
factory-boy==3.3.0
Faker==24.11.0
iniconfig==2.0.0
packaging==24.0
pillow==10.3.0
pluggy==1.5.0
psycopg2-binary==2.9.9
pytest==8.1.1
pytest-django==4.8.0
python-dateutil==2.9.0.post0
setuptools==69.5.1
six==1.16.0
sqlparse==0.5.0
